Plot spatial heatmap

The function SpaHeatMap is used to visualize the physical location vs cluster or batch.

Usage

SpaHeatMap(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

# S3 method for iSCMEBObj
SpaHeatMap(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

# S3 method for iSCMEBResObj
SpaHeatMap(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

# S3 method for list
SpaHeatMap(
  obj,
  item = c("cluster", "batch"),
  plot_type = c("Scatter", "RGB_TSNE", "RGB_UMAP"),
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(1, 2),
  common.legend = TRUE,
  ...
)

Arguments

obj

A iSCMEBObj object or a iSCMEBResObj object.

item

Which feature to be used in the plot.

plot_type

A string shows the type of plot.

title_name

A string shows the title of plot.

combine

An indicator of whether plot all on a figure. If TRUE, all figures are plotted; otherwise, return a list with each plot as component. TRUE by default.

layout.dim

The dimension in the layout of plots when combine = TRUE.

common.legend

An indicator of whether combine the legend of multiple plots. TRUE by default.

...

Arguments passed to other methods.

Value

Return a ggplot2 object or list of ggplots objects.

Return a ggplot2 object or list of ggplots objects.

Return a ggplot2 object or list of ggplots objects.

See also

Examples

data(iSCMEBObj_toy)
#> Warning: data set 'iSCMEBObj_toy' not found
library(Seurat)
p1 <- SpaHeatMap(iSCMEBObj_toy, item="cluster", plot_type="Scatter", common.legend=TRUE)
p2 <- SpaHeatMap(iSCMEBObj_toy, item="cluster", plot_type="RGB_TSNE", common.legend=TRUE)