Fit a iSC.MEB model.

Fit a iSC.MEB model.

Usage

iSCMEB(iSCMEBObj, K = NULL)

Arguments

iSCMEBObj

A iSCMEBObj object.

K

An optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 5~12.

Value

Returns a revised iSCMEBObj object.

Details

The model results are saved in the slot of resList.

See also

Examples

data(iSCMEBObj_toy)
library(Seurat)
iSCMEBObj_toy <- runPCA(iSCMEBObj_toy)
#> Warning: No assay specified, setting assay as RNA by default.
#> Warning: No assay specified, setting assay as RNA by default.
iSCMEBObj_toy <- CreateNeighbors(iSCMEBObj_toy, platform = "Visium")
#> Neighbors were identified for 4222 out of 4222 spots.
#> Neighbors were identified for 4378 out of 4381 spots.
iSCMEBObj_toy <- SetModelParameters(iSCMEBObj_toy, maxIter=10)
iSCMEBObj_toy <- iSCMEB(iSCMEBObj_toy, K = 7)
#> Evaluate initial values...
#> fitting ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> Fit SC-MEB2...
#> Finish variable initialization 
#> K = 7, iter = 2, loglik= -148639.028750, dloglik=0.999931 
#> K = 7, iter = 3, loglik= -142416.378700, dloglik=0.041864 
#> K = 7, iter = 4, loglik= -138046.918112, dloglik=0.030681 
#> K = 7, iter = 5, loglik= -135153.817759, dloglik=0.020957 
#> K = 7, iter = 6, loglik= -133625.320193, dloglik=0.011309 
#> K = 7, iter = 7, loglik= -132511.266520, dloglik=0.008337 
#> K = 7, iter = 8, loglik= -131724.010584, dloglik=0.005941 
#> K = 7, iter = 9, loglik= -131179.790106, dloglik=0.004132 
#> K = 7, iter = 10, loglik= -130786.908993, dloglik=0.002995 
#> K = 7, iter = 11, loglik= -130507.107250, dloglik=0.002139 
#> K = 7, iter = 12, loglik= -130284.749940, dloglik=0.001704 
#> K = 7, iter = 13, loglik= -130105.141717, dloglik=0.001379 
#> K = 7, iter = 14, loglik= -129958.385584, dloglik=0.001128 
#> K = 7, iter = 15, loglik= -129829.964108, dloglik=0.000988 
#> K = 7, iter = 16, loglik= -129721.926764, dloglik=0.000832 
#> K = 7, iter = 17, loglik= -129631.779150, dloglik=0.000695 
#> K = 7, iter = 18, loglik= -129553.580371, dloglik=0.000603 
#> K = 7, iter = 19, loglik= -129486.764811, dloglik=0.000516 
#> K = 7, iter = 20, loglik= -129430.777383, dloglik=0.000432 
#> K = 7, iter = 21, loglik= -129383.961837, dloglik=0.000362 
#> K = 7, iter = 22, loglik= -129339.893631, dloglik=0.000341 
#> K = 7, iter = 23, loglik= -129301.550683, dloglik=0.000296 
#> K = 7, iter = 24, loglik= -129274.973613, dloglik=0.000206 
#> K = 7, iter = 25, loglik= -129243.926110, dloglik=0.000240